requirements.txt
asgiref==3.2.3
boto3==1.12.44
botocore==1.15.44
dj-database-url==0.5.0
Django==3.0.4
django-filter==2.2.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.5
psycopg2==2.8.5
Pillow==7.1.1
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.13.0
sqlparse==0.3.1
urllib3==1.25.7
whitenoise==5.0.1
django-star-ratings==0.9.1
django-bootstrap-form==3.4